Update DataSharing-MS.md
a) not all research is publicly-funded. Some ecological research may be privately funded but granted the % will be very small
b) sometimes the researcher *does* own the data they collect (not that I agree with or endorse this) e.g. for BBSRC funded research: "Ownership of the data generated from the research that BBSRC funds resides with the
investigators and their institutions" (source: http://www.bbsrc.ac.uk/nmsruntime/saveasdialog.aspx?lID=1569&sID=8334 )
 